infrastructure/cdk/requirements.txt
# CDK dependencies
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0

# Python type hints for boto3
boto3-stubs[essential]==1.35.0
